Update try-out-flow.md
The response to the application access token has been changed from the initial response. The scope openId is not returned and the access_token: claim is a jwt. Also the the initial request in the example did not have the payment scope included but the response in the example has the payment scope as well. To correct these points please merge the PR.